Improved handling of attached streams and delayed notifications (#48)
* Mark Context::attach Context::notify_interval and Context::notify_after as must_use

* Downgrade addresses to make sure delayed background notifications don't prevent actors from being dropped

* Cancel attached streams when the `Address` becomes disconnected

* `Address::join` function allows waiting for addresses to finish / become disconnected

* End `Context::notify_interval` and `Context::notify_after` tasks when their `Context` drops

* Removed `stream-cancel` dependency

Removed the `stream-cancel` crate and replaced its functionality with `DropNotice`

* Added missing newline

* Moved `DropNotice` from `Address` to `RefCounter`

It can now be accessed via `disconnect_notice` on `RefCounter`. Also added a dummy `DropNotice` that is used in the `Weak` ref_counter when the `Arc` can't be upgraded.

* Formatting

Let `rustfmt` do its thing

* Test cases added

Added additional tests to cover stream cancellation and `Address::join`

* Lock added to prevent race condition

Added a `RwLock` to `RefCounter` to prevent a possible race
condition that can lead to an Actor not getting stopped.

See
https://github.com/Restioson/xtra/pull/48#issuecomment-983367483
and
https://github.com/Restioson/xtra/pull/48#issuecomment-984304171
for details.

* Clippy related changes

* Added named fields

Made `Strong` and `Weak` structs clearer and added new struct to
`refcount`.