Move `Actor::create` to dedicated trait
This function is not meant to ever be implemented by users.
This is indicated by all the fields in `ActorManager` being
`pub(crate)`. To avoid this confusion, we can move this
functionality to a dedicated trait that lives within the
`manager` module which allows us to make the fields private
altogeher.