Run CI on multiple toolchains
Running CI on beta and nightly is useful for libraries to detect problems in upcoming rust releases. Running on MSRV is useful to know, when we break MSRV and can bump the version appropriately. xtra does not have an explicitly stated policy here so I am just assuming that we want to bump at least a minor version for every MSRV change.