Improve CI coverage and fix doc tests (#69)
* Improve CI scripts

Instead of running on selective features, we run CI with the combination
of `--all-features` and `--no-default-features`. Features are meant to be
additive so this covers most of the cases.

* Cache CI artifacts to improve speed

* Feature-gate `notify_after` correctly

* Run all tests in CI

Only running `--lib` will skip the doc tests.

* Allow `drop_notifier` to be unread if `timing` is not active

* Boolean logic is hard

* Run CI on multiple toolchains

Running CI on beta and nightly is useful for libraries to detect problems in upcoming rust releases. Running on MSRV is useful to know, when we break MSRV and can bump the version appropriately. xtra does not have an explicitly stated policy here so I am just assuming that we want to bump at least a minor version for every MSRV change.

* Remove explicit example build

Examples are built when tests are run, no need to do this separately.

* Fix doc tests

* Remove unnecessary platform matrix variable

* Fix spelling error in action parameter

* Make doc tests pass without default features

* Explain disabled cache on nightly

* Trim down CI configuration

* Respect `matrix.args` for cache artifacts

* Actually make condition work

* Don't run CI on nightly

Without being able to cache artifacts, we would always have to wait on this job, rendering it a bit useless.

Running on beta still gives us ~6 weeks to report problems to the Rust team in case an upcoming changes causes trouble with our codebase.