Drop `MessageChannel` Sync bound
A type being Sync means we can safely pass references of it to other
threads without creating a data race. `MessageChannel`s are really just
`Address`es. If we want to share one of these with another thread, we can
just create a copy and move it over. Thus, the most important requirement
is that they are `Send`.