Only allow "unit" handlers to be `Sink`s
Sending an item into a `Sink` will discard its return value. Allowing any
`Address` to be used as a sink is a foot-gun as the return value could be
something important like a `Result::Err`.

Unfortunately, this means we can no longer guarantee that every
`MessageChannel` is a `Sink`. This isn't particularly problematic though.
`MessageChannel`s are an abstraction to erase the `Actor` type of a `Handler`.
A `Sink` achieves the same thing thus if someone really needs to make a `Sink
out of a `MessageChannel`, they should be able to construct a `Sink` in the
first place and use that instead.