Combine `SendFuture` implementations
We can combine these two implementations by making one that is generic over the
`Sending` future part. In order to provide back-pressure for case of the message
channel, we box up the future which erases the actor's type and thus allows us to
name the return value without mentioning the actor type itself. For the `Address`
we create a dedicated `NameableSending` type. This allows us to avoid an allocation,
making sending of the message cheaper.