Unify the `send` API on `Address` and `MessageChannel` (#74)
* Remove do_send and do_send_async in favor of `recv_async` on `SendFuture`s

* Fix compile error in benchmark

* Store marker type in separate field

* Use `Done` state to represent completion

* Use `Done` instead of `Disconnected` for default

Whilst implementing `Default` is technically not wrong, we don't use it for
anything else other than using `mem::take` in which case, using `mem::replace`
is clearer as it signals that we need a placeholder value.

* Avoid `map_err` in multiple places

By directly storing the mapped receiver, we can avoid calling `map_err`,
thus making the inner state machine easier to understand.

* Avoid allocation for `recv_async`

By introducing a newtype around `catty::Receiver`, we can avoid
the box allocation.

* Align state machines between ReceiveSync and ReceiveAsync

* Rewrite state machine to use mutation

By using mutation, we can make use of recursion to directly trigger
the next branch after we have switched state. This is useful when
we go from `Sending` to `Receiving`.

* Make one constructor per state

* Save a few LoC

* Appease clippy

* Re-order blocks to reduce diff

* Combine `SendFuture` implementations

We can combine these two implementations by making one that is generic over the
`Sending` future part. In order to provide back-pressure for case of the message
channel, we box up the future which erases the actor's type and thus allows us to
name the return value without mentioning the actor type itself. For the `Address`
we create a dedicated `NameableSending` type. This allows us to avoid an allocation,
making sending of the message cheaper.

* Temporarily appease clippy

* Fix clippy-beta lint

* Add tests and example for demonstrating backpressure

* Fix clippy lint

* Move marker types into `send_future` module

* Improve documentation for new APIs

* Remove TODO

The type isn't actually that complex.

* Make test less dependent on executor behaviour

Using `yield_once` gives us precise control over when we handle
the message.

* Run formatter

* Don't break MSRV

* Mark `Receiver` as `must_use`

* Add reason to `must_use` of `SendFuture`

* Typos and doc improvements

* Fix formatting

* Fix benchmark

This used to be an async `do_send`.

* Rename `recv_async` to `split_receiver`

* Run cargo fmt

* Fix compile errors

* Extend docs for `notify_interval` and `notify_after`

* Change old async sends to `split_receiver`