Make sure all examples and benches are compiled in CI
Some examples required additional features in optional dependencies
to be activated. These are not activated with `--all-features` and
thus, these examples were not checked in CI.

To mitigate this, we activate all necessary features as part of our
dev-dependencies which will always be used for all examples and
benchmarks.

As a result, we can check all examples and benchmarks in CI by
activating `--all-features`.