Optimistic heap implementation
Optimised for the case that most priorities are the same - in this way it is optimistic that the user is not mixing and matching priorities a huge amount.

I'm not sure if this is actually better in a real world case, but it should beat std's binary heap for all *existing* Address patterns, as they all use effectively the same priority. This could change in future, though, so std's might actually be better in those cases. This brings us down from 320ns+ to 200ns