Messages priority & inbox unification (#85)
* Super early idea for message priorities

* Optimistic heap implementation

Optimised for the case that most priorities are the same - in this way it is optimistic that the user is not mixing and matching priorities a huge amount.

I'm not sure if this is actually better in a real world case, but it should beat std's binary heap for all *existing* Address patterns, as they all use effectively the same priority. This could change in future, though, so std's might actually be better in those cases. This brings us down from 320ns+ to 200ns

* Remove broadcast queues that have no receivers

* Remove custom binary heap and add default queue

Default queue preserves order unlike the binary heap does. Plus, my implementation of the binary heap was subtly wrong. It would have been fixable, but not really worth it, considering that now that default queue exists, most priority messages probably won't have the same priority (or will be some infrequent that it doesn't matter), so std's strategy will probably beat it out anyway.

* do some work on bounding + fmt

* Tests compile & properly Drop SendFuture

* Code cleanup

* Convert receive to future combinator

This helps handle its drop better (future change)

* Send message back into channel if recv dropped

This solves an edge case

* Allow ReceiveFutures to be cancelled

This should help reduce some out of order receiving when used properly. Theoretically, if `cancel` is used prior to cancelling a `ReceiveFuture`, then out of order messages should never happen. I only expect this to actually become a problem when custom event loops are implemented, or with `select`/`join` being used

* Replace drop_notice infrastructure

* Change eq/ord semantics & add test

* Reimplement fn capacity

* Remove BroadcastMessage, rename wrapper to it

* Fmt

* Address review comments

* Use while let & fmt

* Fix copy-paste error in doc

* Remove redundant map

* Use ? over combinator (shorter)

* Encapsulate ref counters a bit better

* Use map_or

* Rename broadcast / stolen

* pub(crate) -> pub

* TxStrong::new -> TxStrong::try_new

* Remove unnecessary impl of BroadcastEnvelope

* Remove unnecessary Clone impl

* Remove unused AddressMessage type

Superseded by inbox::ActorMessage

* Port tests from #51

* Remove unnecessary early return

* Sink impl

* Fix cloning issue

* Fmt & move broadcast to Address

* Clarify why `message_channel::private` needs to exist (#87)

* Reorder match & remove old comment

* Address some review comments

* Reorder functions

* bounded broadcast POC

* Test other types of send

* Fmt

* impl len

* update tracing

* update async trait

* Update for minimal versions & clippy

* Allow priority to be set

* Add sink test

* Add handle order test

* Add a send waiter order test

* Fmt

* Add broadcast tail test

* Add yet another broadcast tail test

* Fmt

* remove superseded tests

* remove unnecessary `let _`

Thanks beta clippy!

* Introduce `BroadcastFuture` (#88)

* Initial draft of broadcast future

* Add some docs

* Add TODOS

* Add docs for `BroadcastFuture`

* Use `Priority::default` in `BroadcastFuture`

* Use SeqCst where unsure

* Docs

* Docs & remove negative priority messages

* Remove priority enum in favour of option

This is a bit clearer I think

* Fmt

* Set priority for addresses and channels

* update loc count

* Update src/address.rs

Co-authored-by: Thomas Eizinger <thomas@eizinger.io>

* Update src/message_channel.rs

Co-authored-by: Thomas Eizinger <thomas@eizinger.io>

* Update src/context.rs

Co-authored-by: Mariusz Klochowicz <mariusz@klochowicz.com>

* Address review comments & update example in README

* Remove dead comments

Co-authored-by: Thomas Eizinger <thomas@eizinger.io>
Co-authored-by: Mariusz Klochowicz <mariusz@klochowicz.com>