Make `stop_all` behave as `stop_self` on all actors (#113)
Previously, `stop_all` used to drop all messages in the channel and close it immediately. Now, it is equivalent to running `stop_self` on all running actors, and it does not have this special undocumented behaviour. This necessitated a different way of notifying shutdowns, which is why a new shutdown envelope was introduced.

Co-authored-by: Thomas Eizinger <thomas@eizinger.io>