Always store message priority in envelope (#142)
* Always store message priority in envelope

Previously, we only stored the priority of broadcast messages in
the envelope. For regular messages, the priority was stored
outside in a wrapper struct.

This disparity is hard to understand as there is no apparent reason
why the priority could not be stored in the message envelope.

With this patch, we move the priority into the message envelope
and also expose a `set_priority` function for each envelope type
which allows us to consolidate the way we set priority from the
`SendFuture`. To allow this, message envelopes need to be employ
interior mutability (they are either box'ed or arc'ed). We use
a spin lock which should basically be free. A `SendFuture` cannot
be cloned and thus will never be accessed from two different
threads, meaning the lock will never be congested.