Strip span fields from waiting and handler spans
Currently, these are always children of the request span, so adding them as fields in them too is redundant.

It would be a simplification to move this just to xtra_message_handler, but this is unfortunately not possible in the case of a disconnected actor which will never handle the message - this would miss all actor and message type information.