Employ "double result" pattern for `try_send`
Using a double result is useful in cases where the errors returned
by a function fall into two categories: recoverable and non-recoverable.

A disconnected channel is not recoverable and we can thus directly
forward it with `?`. The mailbox being full is something we need to
deal with in `SendFuture` which is why it is within another layer
of `Result`.