Unify message sending into one `SendFuture` (#147)
* Have `SendFuture` handle every case of sending messages

This one future now handles:

- Sending a single message with the actor type being named
- Sending a single message with the actor type being erased
- Sending a broadcast message with the actor type being named

* Add SendFuture::broadcast_erased

Using this function takes a few more changes but it is added here
because it will definitely be useful later.

* Employ "double result" pattern for `try_send`

Using a double result is useful in cases where the errors returned
by a function fall into two categories: recoverable and non-recoverable.

A disconnected channel is not recoverable and we can thus directly
forward it with `?`. The mailbox being full is something we need to
deal with in `SendFuture` which is why it is within another layer
of `Result`.