Encapsulate more functionality in `WaitingReceiver`  (#149)
* Make `WakeReason` an implementation detail of `WaitingReceiver`

We can achieve this by having more specific functions on
`WaitingReceiver` itself.

* Introduce `FulfillHandle` for `WaitingReceiver`

Previously, the references to `WaitingReceiver` were stored "naked"
within the channel and the `Future` implementation. This made it hard
to understand that there are only ever two references.

This patch introduces those two references as proper types and adds
documentation on how the `WaitingReceiver` works.

* Store cancellation state in boolean flag

This avoids an "unreachable" branch when polling the receiver.

* Make `WakeReason` an implementation detail of `WaitingReceiver`

We can achieve this by making a custom `poll` function on
`WaitingReceiver` and passing `Receiver` as a reference in.

* Make `WakeReason` private

This is possible, now that we made it an implementation detail of
`WaitingReceiver`.

* Move `WaitingReceiver` to its own module

This component houses a good chunk of functionality now and is
better understood in isolation.

Additionally, this allows us to better enforce visibility of types.

* Introduce dedicated `Waiting` sub-state

This allows for a more focused `Drop` implementation.

* Use `catty` to implement `WaitingReceiver` (#151)

Now that the waker management is encapsulated in catty, `WakeReason`
is no longer an accurate name. `CtrlMsg` seems better.

Additionally, instead of "fulfilling" the waiting receiver, the
language of "notifying" feels more natural.

* Remove dependency on catty patch (#153)

* Remove dependency on catty patch

Co-authored-by: Restioson <restiosondev@gmail.com>

* Update src/inbox.rs

Co-authored-by: Restioson <restiosondev@gmail.com>