Redesign xtra's spawn API (#156)
* Redesign xtra's spawn API

This removes `Actor::create`, `ActorManager` and the various runtime
extension traits in favor of top-level free functions for spawning
an actor instance into a specific runtime.

This is deemed to be more ergonomic because:

- Users do not need to import an extension trait for their runtime
- Users do not need to import the `Actor` trait to run it which was
otherwise required to call `create`.
- The implementation has less indirection, thus making it easier for
users to understand and perhaps adapt / write their own `spawn` fn
depending on their needs.

The implementation is not strictly functionally equivalent to what
we had before. In particular, we no longer support spawning in a
runtime handle of the tokio and smol runtimes. We accept this
regression because we assume that it is a niche usecase and the now
easier implementation should showcase users, how they can easily
write themselves a helper function that spawns an actor using a
runtime handle in a single line of code. In other words, the `spawn`
API of xtra is purposely design to make 80% of usecases work. For
everything else, users should write their own `spawn` function.

Closes #136.