Simplify control-flow in `ChanInner::pop_` functions
Previously, we first checked whether the corresponding queue WILL
have space if we now pop a message.

This is functionally equivalent to _trying_ to pop a message, exiting
early if we can't (using `?`) and then adding another message to the
queue if we now have space.

Checking for space is equivalent to the queue not being full and we
already have helper functions for this.