Differentiate single and broadcast message on type-system level (#166)
Prior to this patch-set, we had constructs like `MessageType` and `SentMessage`. These were used to write code that is generic over a particular message type. In reality however, we don't actually need to differentiate between these cases because the codepaths for each type are statically known in all cases.

Unfortunately, introducing an actual split introduced a bit more code (net ~ 70 lines) but in exchange, we remove a few `unreachable` error clauses and make it overall easier to follow the dataflow through xtra's channel implementation.